refactor: rename groupe
