+++ components/protected-certificates/qualibat-section.tsx
