+++ components/protected-certificates/qualifelec-section.tsx
